Adding Standalone SAM Model link refs #241 Title capitalized and link to transient model in GitHub added
Title capitalized and link to model on GitHub was added

Title fully capitalized and link to correct folder in GitHub added.

Title fully capitalized and link to GitHub files added.

Link to model folder in GitHub was added under title.

Link to model file in GitHub added under contact.

Title capitalized and link to model in GitHub added.

Title capitalized and link to model in GitHub added.

Link to model in GitHub added.

Link to model in GitHub added.

Link to model in GitHub added.

Title capitalized and link to model in GitHub added.

Link to model in GitHub added.

Link to model in GitHub added.

Italicized the contact line and added link to model in GitHub.

Capitalized title and added link to model in GitHub.

Added link to model in GitHub.

Capitalized title and added link to model in GitHub.

Added link to model in GitHub.

Added link to model in GitHub.

Title capitalized and link added to model in GitHub.

Added link to model in GitHub.

Added link to model in GitHub.

Added link to model in GitHub.

Title capitalized and added link for model in GitHub.

Title capitalized and added link to model in GitHub.

Added link to model in GitHub.

Added link to model in GitHub.

Model link template added to VTB template page and contact line italicized.

Added link to model in GitHub for LFR single assembly model.

Title capitalized, contact line reorganized, and link added for model in GitHub on the mrad multiphysics model page.

Added "legacy" to title and fixed capitalization. Also added link to model in GitHub for legacy HPMR model page.

Added link to model in GitHub for S8ER model page.

Added title to page and link to model in GitHub for GCMR assembly model.