Updated deprecated input, re-golded
Changed HeatStructureMaterials block to SolidProperties block, turned off the
use of deprecated volume junction scalar variables, changed parameter name
‘function’ to ‘expression’ in ParsedPostprocessors. Updated other parts of
input to be compatible with these changes. The updated input had the same
results in the csv file as the old version, with some additional columns.
A re-gold was performed to reflect this change.