Removed extra semicolon from input.
MOOSE has recently been updated so that, when parsing a semicolon-seperated
vector of vectors from an input file, a trailing semicolon on the last vector
will result in the creation of an additional vector of size zero. This effect
is not intended for the gFHR_griffin_cr_ss.i input, so the trailing semicolon
has been removed.