Changed FileMesh object to FileMeshGenerator object
The object that reads the mesh for the microreactors/gcmr/assembly/steady_state.bison_alone_transient_syntax
test is specified in the cli_args as a FileMesh. However, the input file contains
a FileMeshGenerator, which causes conflicts and ultimately, a test crash. This
commit changes the FileMesh in the cli_args to a FileMeshGenerator to be consistent
with what is provided in the input file.

Ref #526